Move ConnectProxyConnector to its own module
